optimize the Words component
I tried to refactor the big block in handleKey

only the line that changes on a key press will be reconstructed, rather than reconstructing every line on every render

i still feel i can optimize this further, but this will do for now.
i may try to optimize this again in the future.