fix appearance animation for incorrect characters
i broke this in f23b320. i forgot that i needed the inline-block for that animation because thats required to animate the width of a span element. since every span element is no longer a block element, using inline-block on only the incorrect character caused some alignment issues. inline-flex seems to be fine, so that's why i'm using it here.